requirements.txt
# Core framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic-settings==2.7.1
python-multipart==0.0.20
orjson==3.10.14

# ML / Face recognition
facenet-pytorch==2.6.0
torch==2.5.1
torchvision==0.20.1
pillow==11.1.0
numpy==2.2.2

# Database
sqlalchemy==2.0.36
psycopg[binary]==3.2.4
pgvector==0.3.6
alembic==1.14.1

# Observability & utilities
structlog==24.4.0
httpx==0.28.1
prometheus-client==0.21.1

# Security
python-jose[cryptography]==3.3.0
